GuiImpl::RunMsg: fn_diff 100% — full msg interpreter match
- Frame layout via #pragma var_order (args/i15/args15/i17/args17/bufN/i9)
- Case 15/17: showPortrait fields (i32 anmScriptIdx0 for case 17 SetSprite)
- Cases 16/19/20: DecryptMsgText src = args+0 (ZUN quirk, not args+4)
- Case 21/4: inverted if + goto shared break label for MSVC conditional-jump layout
- Case 7: inline musicIdx, PlayMusic + conditional PlayAudio via 0x4c7240 table
- Case 9: inline stageResults addressing, i8 GetClockTime, shotType bombs check
- GuiMsgState framesElapsedDuringPause/pauseLimit are i32

+++ src/GameManager.cpp
@@ -906,9 +906,9 @@ void GameManager::CutChain()
 }
 
 // FUNCTION: th08 0x406dd0
-u8 GameManager::GetClockTime()
+i8 GameManager::GetClockTime()
 {
-    return this->globals->clockTime;
+    return (i8)this->globals->clockTime;
 }
 
 // FUNCTION: th08 0x406df0
